git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@11308 01821946-851b-0410-b7f4-f1477c70e57a

+++ network/network_send.cpp
@@ -170,14 +170,14 @@ void TestXMLInfoRequest (ubyte* serverAddress)
 {
 #if 0 //DBG
 gameStates.multi.bTrackerCall = 2;
-IPXSendInternetPacketData((ubyte *) "FDescent Game Info Request", strlen ("FDescent Game Info Request") + 1, 
-								  serverAddress, serverAddress + 4);
+IPXSendInternetPacketData ((ubyte *) "FDescent Game Info Request", (int) strlen ("FDescent Game Info Request") + 1, 
+								   serverAddress, serverAddress + 4);
 gameStates.multi.bTrackerCall = 0;
 #endif
 #if 1 //DBG
 gameStates.multi.bTrackerCall = 2;
-IPXSendInternetPacketData((ubyte *) "GDescent Game Status Request", strlen ("GDescent Game Status Request") + 1, 
-								  serverAddress, serverAddress + 4);
+IPXSendInternetPacketData ((ubyte *) "GDescent Game Status Request", (int) strlen ("GDescent Game Status Request") + 1, 
+								   serverAddress, serverAddress + 4);
 gameStates.multi.bTrackerCall = 0;
 #endif
 }
